requirements.txt
beautifulsoup4==4.12.3
Flask==3.0.3
kafka-python==2.0.2
requests==2.32.3
